IGNITE-23488 Integrate cancellation of transaction with sql kill handler.
