+++ modules/compute/src/main/java/org/apache/ignite/internal/compute/IgniteComputeImpl.java
@@ -67,8 +67,6 @@ import org.apache.ignite.internal.hlc.HybridClock;
 import org.apache.ignite.internal.network.TopologyService;
 import org.apache.ignite.internal.placementdriver.PlacementDriver;
 import org.apache.ignite.internal.replicator.TablePartitionId;
-import org.apache.ignite.internal.sql.engine.api.kill.CancellableOperationType;
-import org.apache.ignite.internal.sql.engine.api.kill.OperationKillHandler;
 import org.apache.ignite.internal.table.IgniteTablesInternal;
 import org.apache.ignite.internal.table.StreamerReceiverRunner;
 import org.apache.ignite.internal.table.TableViewInternal;
@@ -614,29 +612,6 @@ public class IgniteComputeImpl implements IgniteComputeInternal, StreamerReceive
 
                     return SharedComputeUtils.unmarshalArgOrResult(res, null, null);
                 });
-    }
-
-    /** Returns a {@link OperationKillHandler kill handler} for the compute job. */
-    public OperationKillHandler killHandler() {
-        return new OperationKillHandler() {
-            @Override
-            public CompletableFuture<Boolean> cancelAsync(String operationId) {
-                UUID jobId = UUID.fromString(operationId);
-
-                return IgniteComputeImpl.this.cancelAsync(jobId)
-                        .thenApply(res -> res != null ? res : Boolean.FALSE);
-            }
-
-            @Override
-            public boolean local() {
-                return false;
-            }
-
-            @Override
-            public CancellableOperationType type() {
-                return CancellableOperationType.COMPUTE;
-            }
-        };
     }
 
     @TestOnly
