+++ modules/sql-engine/src/integrationTest/java/org/apache/ignite/internal/sql/engine/kill/ItSqlKillCommandTest.java
@@ -35,7 +35,6 @@ import static org.hamcrest.Matchers.instanceOf;
 import static org.hamcrest.Matchers.is;
 import static org.hamcrest.Matchers.not;
 import static org.hamcrest.Matchers.nullValue;
-import static org.junit.jupiter.api.Assumptions.assumeTrue;
 
 import java.util.Collection;
 import java.util.List;
@@ -104,9 +103,6 @@ public class ItSqlKillCommandTest extends BaseSqlIntegrationTest {
     @ParameterizedTest
     @EnumSource(CancellableOperationType.class)
     public void killWithInvalidIdentifier(CancellableOperationType type) {
-        // TODO https://issues.apache.org/jira/browse/IGNITE-23488 Remove assumption.
-        assumeTrue(type != CancellableOperationType.TRANSACTION, type + " not implemented yet");
-
         Consumer<String> exceptionChecker = query -> {
             SqlException err = assertThrowsSqlException(
                     SqlException.class,
@@ -126,9 +122,6 @@ public class ItSqlKillCommandTest extends BaseSqlIntegrationTest {
     @ParameterizedTest
     @EnumSource(CancellableOperationType.class)
     public void killNonExistentOperation(CancellableOperationType type) {
-        // TODO https://issues.apache.org/jira/browse/IGNITE-23488 Remove assumption.
-        assumeTrue(type != CancellableOperationType.TRANSACTION, type + " not implemented yet");
-
         assertThat(executeKill(CLUSTER.aliveNode(), type, UUID.randomUUID(), false), is(false));
 
         // NO WAIT should never return false.
