+++ modules/transactions/src/main/java/org/apache/ignite/internal/tx/impl/TxManagerImpl.java
@@ -870,7 +870,7 @@ public class TxManagerImpl implements TxManager, NetworkMessageHandler, SystemVi
 
             txStateVolatileStorage.start();
 
-            txViewProvider.init(localNodeId, lowWatermark.lockIds(), txStateVolatileStorage.statesMap());
+            txViewProvider.init(localNodeId, txStateVolatileStorage.statesMap());
 
             orphanDetector.start(txStateVolatileStorage, txConfig.abandonedCheckTs());
